Rearrange code slightly
Making it closer to Prismatic's guidelines, just because I'm more
comfortable thinking in that sort of namespace layout. Bigger change was
adding schema to make it easier for me to track which parameters are
what.

Started updating the underlying atom as a first step in applying
rules (so now it has a field to track whose turn it is)