Seems to be the tipping point
If I have draw-tuple wrap its pair of squares in a span, life gets
crazy. I had to do this when I was trying to treat it as a Component.

If it just returns a pair of td elements that get mapcat'd so that
draw-row's tr can apply them, everything seems fine.