Lots of debugging
It all boiled down to forgetting that a component has to return a single
element. The results of build-all don't cut it.

This version at least draws my board...but it's getting mounted and then
unmounted pretty much immediately. It also has that same warning about
invalid checksums and the server rendering that's been giving me so many fits.